Reorganized some DTO's into folders
