+++ src/Dating App/3. BLL/DatingApp.BLL/DTO/AppUser/AppUserLoginDto.cs
@@ -1,12 +1,12 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace DatingApp.BLL.DTO
+namespace DatingApp.BLL.DTO.AppUser
 {
     public class AppUserLoginDto
     {
         [Required]
         public string Username { get; set; }
-        
+
         [Required]
         public string Password { get; set; }
